+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 
+- Log Redis command arguments when sending PII is enabled [#1726](https://github.com/getsentry/sentry-ruby/pull/1726)
 - Add request env to sampling context [#1749](https://github.com/getsentry/sentry-ruby/pull/1749)
 
 ### Bug Fixes