+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## Unreleased
+
+### Features
+
+- Log Redis command arguments when sending PII is enabled [#1726](https://github.com/getsentry/sentry-ruby/pull/1726)
+- Add request env to sampling context [#1749](https://github.com/getsentry/sentry-ruby/pull/1749)
+
+### Bug Fixes
+
+- Require set library [#1753](https://github.com/getsentry/sentry-ruby/pull/1753)
+  - Fixes [#1752](https://github.com/getsentry/sentry-ruby/issues/1752)
+
 ## 5.1.1
 
 ### Bug Fixes